Turn DynamicComparer private class into public class in Types namespace
